+++ src/main/java/me/xpyex/plugin/flywithfood/common/implementations/FWFInfo.java
@@ -4,21 +4,21 @@ import me.xpyex.plugin.flywithfood.common.implementations.flyenergy.EnergyManage
 import me.xpyex.plugin.flywithfood.common.implementations.flyenergy.FlyEnergy;
 
 public class FWFInfo {
-    private final double cost;
-    private final double disable;
+    private final Integer cost;
+    private final Integer disable;
     private final FlyEnergy energy;
 
-    public FWFInfo(double cost, double disable, String energy) {
+    public FWFInfo(Integer cost, Integer disable, String energy) {
         this.cost = cost;
         this.disable = disable;
         this.energy = EnergyManager.getEnergy(energy);
     }
 
-    public double getCost() {
+    public Integer getCost() {
         return this.cost;
     }
 
-    public double getDisable() {
+    public Integer getDisable() {
         return this.disable;
     }
